Moved http calls out into spotify.http.service to help improve testability and separation of concerns
