+++ src/services/spotify/spotify.service.ts
@@ -1,43 +1,26 @@
-import axios, { AxiosResponse } from 'axios';
+import { AxiosResponse } from 'axios';
 import { isWithinInterval, subHours } from 'date-fns';
 import { Playlist } from '../../shared/db/models/Playlist';
 import { Song } from '../../shared/db/models/Song';
 import { User } from '../../shared/db/models/User';
 import { QueueService } from '../../shared/services/queue.service';
 import { UserService } from '../user/user.service';
+import { SpotifyHttpService } from './spotify-http.service';
 import {
-  SpotifyLikedSong,
   SpotifyPlaylist,
   SpotifyPlaylistItemInfo,
   SpotifyResponse,
-  SpotifyTrack,
   SpotifyUserData,
 } from './spotify.generated.interface';
 import { PlaylistData, SongsByUser, SongWithUserData } from './spotify.interface';
 
 export class SpotifyService {
-  baseUrl = 'https://api.spotify.com/v1';
-  baseSelfUrl = `${this.baseUrl}/me`;
-  baseUserUrl = `${this.baseUrl}/users`;
-  basePlaylistUrl = `${this.baseUrl}/playlists`;
-
   userService = new UserService();
   queueService = QueueService.getInstance();
+  httpService = new SpotifyHttpService();
 
   getUserData(accessToken: string): Promise<SpotifyUserData> {
-    return axios
-      .get(this.baseSelfUrl, {
-        headers: {
-          Authorization: 'Bearer ' + accessToken,
-        },
-      })
-      .then(response => {
-        return response.data;
-      })
-      .catch(e => {
-        console.error(e);
-        throw new Error(e);
-      });
+    return this.httpService.getUserData(accessToken);
   }
 
   async getUserPlaylists(accessToken: string): Promise<PlaylistData> {
@@ -45,81 +28,40 @@ export class SpotifyService {
       where: { accessToken },
       relations: ['memberPlaylists', 'ownedPlaylists'],
     });
-    return axios
-      .get<SpotifyResponse<SpotifyPlaylist[]>>(`${this.baseSelfUrl}/playlists`, {
-        headers: {
-          Authorization: `Bearer ${accessToken}`,
-        },
-      })
-      .then(
-        (resp: AxiosResponse<SpotifyResponse<SpotifyPlaylist[]>>): PlaylistData => {
-          if (resp) {
-            const spotifyPlaylists: SpotifyPlaylist[] = resp.data.items;
-            if (user) {
-              const createdPlaylists = user[0]?.ownedPlaylists?.map(x => x.playlistId) || [];
-              const memberPlaylists =
-                user[0]?.memberPlaylists?.map(x => x.playlistId)?.filter(x => !createdPlaylists?.includes(x)) || [];
+    return this.httpService.getUserPlaylists(accessToken).then(
+      (resp: AxiosResponse<SpotifyResponse<SpotifyPlaylist[]>>): PlaylistData => {
+        if (resp) {
+          const spotifyPlaylists: SpotifyPlaylist[] = resp.data.items;
+          if (user) {
+            const createdPlaylists = user[0]?.ownedPlaylists?.map(x => x.playlistId) || [];
+            const memberPlaylists =
+              user[0]?.memberPlaylists?.map(x => x.playlistId)?.filter(x => !createdPlaylists?.includes(x)) || [];
 
-              const orphanPlaylists = createdPlaylists.filter(x => !spotifyPlaylists.find(y => x === y.id));
-              const ownedPlaylists = spotifyPlaylists.filter(x => createdPlaylists?.includes(x.id));
-              const subscribedPlaylists = spotifyPlaylists.filter(x => memberPlaylists?.includes(x.id));
+            const orphanPlaylists = createdPlaylists.filter(x => !spotifyPlaylists.find(y => x === y.id));
+            const ownedPlaylists = spotifyPlaylists.filter(x => createdPlaylists?.includes(x.id));
+            const subscribedPlaylists = spotifyPlaylists.filter(x => memberPlaylists?.includes(x.id));
 
-              return {
-                ownedPlaylists,
-                orphanPlaylists,
-                subscribedPlaylists,
-              };
-            }
-            return { ownedPlaylists: [], orphanPlaylists: [], subscribedPlaylists: [] };
+            return {
+              ownedPlaylists,
+              orphanPlaylists,
+              subscribedPlaylists,
+            };
           }
           return { ownedPlaylists: [], orphanPlaylists: [], subscribedPlaylists: [] };
-        },
-      )
-      .catch(e => {
-        console.log(e);
-        throw new Error(e);
-      });
+        }
+        return { ownedPlaylists: [], orphanPlaylists: [], subscribedPlaylists: [] };
+      },
+    );
   }
 
   createUserPlaylist(accessToken: string): Promise<any> {
     return this.userService.getUser({ accessToken: accessToken.split(' ')[1] }).then(user => {
       if (user) {
-        return axios
-          .post(
-            `${this.baseUserUrl}/${user.spotifyId}/playlists`,
-            {
-              name: `${user.spotifyId}'s Remix Playlist`,
-              public: false,
-              collaborative: true,
-              description: 'Playlist generated by SpotifyRemix',
-            },
-            {
-              headers: {
-                Authorization: accessToken,
-              },
-            },
-          )
-          .then(playlist => {
-            return axios
-              .put(
-                `${this.basePlaylistUrl}/${playlist.data.id}`,
-                {
-                  description: `Subscribe and get info about this playlist at ${process.env.SPOTIFY_REMIX_BASE_URL}/playlist?playlistId=${playlist.data.id}`,
-                },
-                {
-                  headers: {
-                    Authorization: accessToken,
-                    'Content-Type': 'application/json',
-                  },
-                },
-              )
-              .then(_ => playlist);
-          })
-          .then(playlist => {
-            return this.userService
-              .savePlaylist(user, playlist.data.id)
-              .then(playlist => this.refreshPlaylist(playlist.playlistId));
-          });
+        this.httpService.createUserPlaylist(accessToken, user).then(playlist => {
+          return this.userService
+            .savePlaylist(user, playlist.data.id)
+            .then(playlist => this.refreshPlaylist(playlist.playlistId));
+        });
       } else {
         throw new Error('Unable to find user');
       }
@@ -151,23 +93,9 @@ export class SpotifyService {
       if (isUserAlreadyMember) {
         return undefined;
       } else {
-        return await axios
-          .put(
-            `${this.basePlaylistUrl}/${playlistId}/followers`,
-            {
-              public: true,
-            },
-            {
-              headers: {
-                Authorization: `Bearer ${accessToken}`,
-              },
-            },
-          )
-          .then(_ => this.userService.updatePlaylistMembers(user[0], playlist[0]))
-          .catch(e => {
-            console.error(e);
-            throw new Error(e);
-          });
+        return this.subscribeToPlaylist(accessToken, playlistId).then(_ =>
+          this.userService.updatePlaylistMembers(user[0], playlist[0]),
+        );
       }
     }
     return undefined;
@@ -175,7 +103,7 @@ export class SpotifyService {
 
   async getTopSongs(members: User[]): Promise<SongsByUser[]> {
     return members?.length
-      ? await Promise.all(members.map(async (member: User) => this.getTopSongsByUser(member)))
+      ? await Promise.all(members.map(async (member: User) => this.httpService.getTopSongsByUser(member)))
       : [];
   }
 
@@ -201,7 +129,7 @@ export class SpotifyService {
       allMusic.map(async (x: SongsByUser) => {
         if (x.topSongs.length < songsPerUser) {
           const newSongsByUser: SongsByUser = Object.assign({}, x);
-          return this.getLikedSongsByUser(x.user).then(y => {
+          return this.httpService.getLikedSongsByUser(x.user).then(y => {
             newSongsByUser.likedSongs = y?.likedSongs?.filter(x => !historyAsStrings.includes(x.uri)) || [];
             return newSongsByUser;
           });
@@ -253,32 +181,9 @@ export class SpotifyService {
     return playlistSongs;
   }
 
-  getPlaylistTracks(playlistId: string, accessToken: string): Promise<SpotifyPlaylistItemInfo[]> {
-    return axios
-      .get<SpotifyResponse<SpotifyPlaylistItemInfo[]>>(`${this.basePlaylistUrl}/${playlistId}/tracks?limit=50`, {
-        headers: {
-          Authorization: `Bearer ${accessToken}`,
-        },
-      })
-      .then((x: AxiosResponse<SpotifyResponse<SpotifyPlaylistItemInfo[]>>): SpotifyPlaylistItemInfo[] => x.data.items);
-  }
-
-  removeAllPlaylistTracks(playlistId: string, accessToken: string, tracks: SpotifyPlaylistItemInfo[]): any {
-    return axios.delete(`${this.basePlaylistUrl}/${playlistId}/tracks`, {
-      headers: {
-        Authorization: `Bearer ${accessToken}`,
-      },
-      data: {
-        tracks: tracks.map(x => ({
-          uri: x.track.uri,
-        })),
-      },
-    });
-  }
-
   refreshPlaylist(playlistId: string): Promise<void> {
     const identifier = `playlist-${playlistId}`;
-    const queue = this.queueService.queue<Playlist[]>(identifier, () => this.populatePlaylist(playlistId));
+    const queue = this.queueService.queue<Playlist>(identifier, () => this.populatePlaylist(playlistId));
     if (queue.length === 1) {
       return this.queueService.dequeue(identifier);
     } else {
@@ -288,7 +193,7 @@ export class SpotifyService {
     }
   }
 
-  private async populatePlaylist(playlistId: string): Promise<Playlist[]> {
+  private async populatePlaylist(playlistId: string): Promise<Playlist> {
     const playlist = await this.userService.getPlaylist(playlistId).then(playlist => {
       return playlist[0];
     });
@@ -298,23 +203,15 @@ export class SpotifyService {
     const music: SongWithUserData[] = await this.getAllMusic(members, songsPerUser, history);
     const orderedPlaylist: SongWithUserData[] = this.roundRobinSort(music);
     // Get all songs from the playlist.
-    const playlistTracks: SpotifyPlaylistItemInfo[] = await this.getPlaylistTracks(playlistId, owner.accessToken);
+    const playlistTracks: SpotifyPlaylistItemInfo[] = await this.httpService.getPlaylistTracks(
+      playlistId,
+      owner.accessToken,
+    );
     // Remove all songs from the playlist.
-    await this.removeAllPlaylistTracks(playlistId, owner.accessToken, playlistTracks);
-    return axios
-      .post(
-        `${this.basePlaylistUrl}/${playlistId}/tracks`,
-        {
-          uris: orderedPlaylist.map(song => song.uri),
-        },
-        {
-          headers: {
-            Authorization: `Bearer ${owner.accessToken}`,
-          },
-        },
-      )
-      .then(_ => this.userService.saveSongs(playlist, orderedPlaylist))
-      .catch(e => e);
+    await this.httpService.removeAllPlaylistTracks(playlistId, owner.accessToken, playlistTracks);
+    return this.httpService
+      .addSongsToPlaylist(owner.accessToken, playlistId, orderedPlaylist)
+      .then(_ => this.userService.saveSongs(playlist, orderedPlaylist));
   }
 
   // Note: This sort is a "best effort" to maintain order within the playlist.
@@ -343,72 +240,6 @@ export class SpotifyService {
     const songsPerUser = numberOfUsers * 6;
     const maxNumberOfSongs = 48;
     return songsPerUser > maxNumberOfSongs ? minSongsPerUser : Math.round(maxNumberOfSongs / numberOfUsers);
-  }
-
-  private getTopSongsByUser(
-    user: User,
-    url = `${this.baseSelfUrl}/top/tracks?limit=50&time_range=short_term`,
-  ): Promise<SongsByUser> {
-    return axios
-      .get<SpotifyResponse<SpotifyTrack[]>>(url, {
-        headers: {
-          Authorization: `Bearer ${user.accessToken}`,
-        },
-      })
-      .then<SongsByUser>(
-        (x: AxiosResponse<SpotifyResponse<SpotifyTrack[]>>): Promise<SongsByUser> => {
-          const songs = x.data.items.map(
-            (song: SpotifyTrack): SongWithUserData => ({
-              ...song,
-              spotifyId: user.spotifyId,
-            }),
-          );
-
-          if (x.data.next) {
-            return this.getTopSongsByUser(user, x.data.next).then((data: SongsByUser) => ({
-              user: user,
-              topSongs: songs.concat(data.topSongs),
-            }));
-          }
-          return new Promise((resolve, _reject) => resolve({ user, topSongs: songs }));
-        },
-      )
-      .catch(e => {
-        console.error(e);
-        throw new Error(e);
-      });
-  }
-
-  private getLikedSongsByUser(user: User, url = `${this.baseSelfUrl}/tracks?limit=50`): Promise<SongsByUser> {
-    return axios
-      .get<SpotifyResponse<SpotifyLikedSong[]>>(url, {
-        headers: {
-          Authorization: `Bearer ${user.accessToken}`,
-        },
-      })
-      .then<SongsByUser>(
-        (x: AxiosResponse<SpotifyResponse<SpotifyLikedSong[]>>): Promise<SongsByUser> => {
-          const songs = x.data.items.map(
-            (song: SpotifyLikedSong): SongWithUserData =>
-              Object.assign(song.track, {
-                spotifyId: user.spotifyId,
-              }),
-          );
-
-          if (x.data.next) {
-            return this.getLikedSongsByUser(user, x.data.next).then(data => ({
-              user,
-              topSongs: [], // Dont love this.
-              likedSongs: songs.concat(data.likedSongs || []),
-            }));
-          }
-          return new Promise((resolve, _reject) => resolve({ user, topSongs: [], likedSongs: songs }));
-        },
-      )
-      .catch(e => {
-        console.error(e);
-        throw new Error(e);
-      });
   }
 
   getPlaylistHistory(playlistId: string): Promise<Song[]> {
